Move settings to a folder - use compiled prefabs.
