+++ src/splitguides/livesplit_client/__init__.py
@@ -1,5 +1,8 @@
-from .messaging import LivesplitMessaging
-from .networking import LivesplitConnection
+from prefab_classes import prefab_compiler
+
+with prefab_compiler():
+    from .messaging import LivesplitMessaging
+    from .networking import LivesplitConnection
 
 
 def get_client(server="localhost", port=16834, timeout=1):
